pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jointed"
version = "0.1.0"
description = "Word-tag puzzle API"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.115.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
